test: move basic tests into own file
