+++ tests/basic.test.ts
@@ -1,0 +1,18 @@
+import { test, expect } from "vitest";
+import { Parser } from "@/Parser";
+import { Tokenizer } from "@/Tokenizer";
+
+test("returns an empty Program", () => {
+  // GIVEN
+  const program = "";
+  const subject = new Parser(new Tokenizer());
+
+  // WHEN
+  const result = subject.parse(program);
+
+  // THEN
+  expect(result).toStrictEqual({
+    type: "Program",
+    body: [],
+  });
+});
